[Engine] [Animation] commit skin animation

+++ Engine/Assets/Engine/Shaders/Forward/Standard_Skin.hlsl
@@ -42,10 +42,10 @@ Pixel mainVS(Vertex_PosUvNorTanBone input)
 
     float4 weights = float4(input.boneWeights.x, input.boneWeights.y, input.boneWeights.z,
     1.0f - input.boneWeights.x - input.boneWeights.y - input.boneWeights.z);
-    matrix boneTransform = bone_data_arr.boneTransformArr[input.boneIndices[0]] * weights[0];
-    boneTransform += bone_data_arr.boneTransformArr[input.boneIndices[1]] * weights[1];
-    boneTransform += bone_data_arr.boneTransformArr[input.boneIndices[2]] * weights[2];
-    boneTransform += bone_data_arr.boneTransformArr[input.boneIndices[3]] * weights[3];
+    matrix boneTransform = bone_data_arr.boneTransformArr[(int) input.boneIndices[0]] * weights[0];
+    boneTransform += bone_data_arr.boneTransformArr[(int) input.boneIndices[1]] * weights[1];
+    boneTransform += bone_data_arr.boneTransformArr[(int) input.boneIndices[2]] * weights[2];
+    boneTransform += bone_data_arr.boneTransformArr[(int) input.boneIndices[3]] * weights[3];
     
     input.position.w = 1.0f;
     float4 posL = mul(input.position, boneTransform);
